[HttpKernel] [Kernel] Fix deprecation notices for cache clearer and warmer
