+++ src/Symfony/Component/Kernel/CacheClearer/ChainCacheClearer.php
@@ -18,22 +18,6 @@ namespace Symfony\Component\Kernel\CacheClearer;
  *
  * @final
  */
-class ChainCacheClearer implements CacheClearerInterface
+class ChainCacheClearer extends InternalChainCacheClearer
 {
-    private $clearers;
-
-    public function __construct(iterable $clearers = [])
-    {
-        $this->clearers = $clearers;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function clear($cacheDir)
-    {
-        foreach ($this->clearers as $clearer) {
-            $clearer->clear($cacheDir);
-        }
-    }
 }
