+++ src/Symfony/Component/Kernel/CacheClearer/InternalChainCacheClearer.php
@@ -1,0 +1,39 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Kernel\CacheClearer;
+
+/**
+ * ChainCacheClearer.
+ *
+ * @author Dustin Dobervich <ddobervich@gmail.com>
+ *
+ * @internal
+ */
+class InternalChainCacheClearer implements CacheClearerInterface
+{
+    private $clearers;
+
+    public function __construct(iterable $clearers = [])
+    {
+        $this->clearers = $clearers;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function clear($cacheDir)
+    {
+        foreach ($this->clearers as $clearer) {
+            $clearer->clear($cacheDir);
+        }
+    }
+}
